fix: cast BookingPrice to int in bill queries and update BillResponse to use nullable int for BookPrice

+++ RestaurantManagement.Application/Features/BillFeature/Queries/GetAllBill/GetAllBillQueryHandler.cs
@@ -74,7 +74,7 @@ public class GetAllBillQueryHandler(
                 a.BillId,
                 a.CreatedDate,
                 a.Booking.BookId,
-                a.Booking.BookingPrice/2,
+                (int)a.Booking.BookingPrice/2,
                 a.Booking.BookingDate,
                 a.Booking.BookingTime,
                 a.Order.OrderId,
